oscillo voice data mngt improvements & bug fixes

+++ src/ModizerVoicesData.h
@@ -9,6 +9,7 @@
 #define ModizerVoicesData_h
 
 #include "ModizerConstants.h"
+#include <sys/types.h>
 
 extern signed char *m_voice_buff[SOUND_MAXVOICES_BUFFER_FX];
 extern signed int *m_voice_buff_accumul_temp[SOUND_MAXVOICES_BUFFER_FX];
@@ -16,8 +17,8 @@ extern unsigned char *m_voice_buff_accumul_temp_cnt[SOUND_MAXVOICES_BUFFER_FX];
 
 extern int m_genNumVoicesChannels;
 
-extern int m_voice_current_ptr[SOUND_MAXVOICES_BUFFER_FX];
-extern int m_voice_prev_current_ptr[SOUND_MAXVOICES_BUFFER_FX];
+extern int64_t m_voice_current_ptr[SOUND_MAXVOICES_BUFFER_FX];
+extern int64_t m_voice_prev_current_ptr[SOUND_MAXVOICES_BUFFER_FX];
 extern int m_voice_ChipID[SOUND_MAXVOICES_BUFFER_FX];
 extern int m_voice_systemColor[SOUND_VOICES_MAX_ACTIVE_CHIPS];
 extern int m_voice_voiceColor[SOUND_MAXVOICES_BUFFER_FX];
